refactor: review 패키지 구조 및 파일 위치 수정
