+++ src/main/java/com/idea5/four_cut_photos_map/domain/member/dto/response/MemberDto.java
@@ -1,4 +1,4 @@
-package com.idea5.four_cut_photos_map.domain.review.dto.entity;
+package com.idea5.four_cut_photos_map.domain.member.dto.response;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
@@ -12,8 +12,7 @@ import lombok.*;
 @NoArgsConstructor
 @ToString
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class MemberResp {
+public class MemberDto {
     private Long id;            // 회원 번호
-    private String nickname;    // 닉네임
-    private String mainMemberTitle; // 회원 대표 칭호
+    private String nickname;    // 카카오 닉네임
 }
