+++ src/main/java/com/idea5/four_cut_photos_map/domain/shop/dto/response/ShopDto.java
@@ -1,5 +1,4 @@
-package com.idea5.four_cut_photos_map.domain.review.dto.entity;
-
+package com.idea5.four_cut_photos_map.domain.shop.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -12,7 +11,8 @@ import lombok.*;
 @NoArgsConstructor
 @ToString
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class MemberDto {
-    private Long id;            // 회원 번호
-    private String nickname;    // 카카오 닉네임
+public class ShopDto {
+    private Long id;            // 가게 번호
+    private String brand;       // 브랜드 명
+    private String placeName;        // 가게 이름
 }
